refactor(collator): use common api to spawn validation cancel
