+++ collator/src/manager/mod.rs
@@ -425,21 +425,6 @@ where
         Ok(())
     }
 
-    // TODO: move to cancel_validation_runner.rs
-    #[tracing::instrument(skip_all, fields(block_id = %state.block_id().as_short_id()))]
-    fn cancel_validation_sessions_until_block(&self, state: ShardStateStuff) -> Result<()> {
-        let block_id = state.block_id().as_short_id();
-
-        let session_id = self
-            .active_collation_sessions
-            .read()
-            .get(&block_id.shard)
-            .map(|session_info| session_info.get_validation_session_id());
-
-        self.validator.cancel_validation(&block_id, session_id)?;
-        Ok(())
-    }
-
     #[tracing::instrument(skip_all, fields(block_id = %block_id))]
     fn commit_block_queue_diff(
         mq_adapter: Arc<dyn MessageQueueAdapter<EnqueuedMessage>>,
@@ -1197,15 +1182,7 @@ where
         let is_key_block = ctx.state.state_extra()?.after_key_block;
 
         // stop any running validations up to this block
-        // try to get the validation session ID from active collation sessions
-        let session_id = self
-            .active_collation_sessions
-            .read()
-            .get(&block_id.shard)
-            .map(|session_info| session_info.get_validation_session_id());
-
-        self.validator
-            .cancel_validation(&block_id.as_short_id(), session_id)?;
+        self.schedule_cancel_validation_sessions_until_block(ctx.state.clone());
 
         // check if should sync to last applied mc block right now
         let should_sync_to_last_applied_mc_block = 'check_should_sync: {
